--- src/Toolbar/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "toolbar",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./Toolbar.js"
-}
